lessons/09-database-services/src/cosmos-test-app/requirements.txt
azure-cosmos==4.7.0
